Update of algo1 and frontend.
I fixed the issues with the consecutive_days input in the dash-app. In the next commit, we need to change/insert a consecutive_days_sell argument, so that we sell as soon/from day 1 the sell signals are met. In that way we don't wait. This could also be a variable, so that we can change it if we want to be more confident in our short positions.